requirements.txt
Pillow
numpy
GitPython